% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cpgInfo}
\alias{cpgInfo}
\title{nearestGenes}
\usage{
cpgInfo(cpgs, TxDb)
}
\arguments{
\item{cpgs}{list of illumina cpg identifiers}

\item{TxDb}{name of Transcript database, i.e. TxDb.Hsapiens.UCSC.hg19.knownGene}
}
\value{
data.frame with cpg annotations
}
\description{
find gene nearest to list of cpgs
}
\details{
find gene nearest to list of cpgs
}
\author{
mvaniterson
}
