% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detectionP}
\alias{detectionP}
\title{detectionP that accepts NA's}
\usage{
detectionP(rgSet, type = "m+u", na.rm = FALSE)
}
\arguments{
\item{rgSet}{RGset}

\item{type}{"m+u"}

\item{na.rm}{FALSE/TRUE}
}
\value{
matrix with detection P-values
}
\description{
detectionP that accepts NA's
}
\details{
since probe-filtering can introduce NA's
detectionP is modified to handle NA's properly
beware the P-value matrix thus can contain NA's
}
\author{
mvaniterson
}
