% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{DMRfinder}
\alias{DMRfinder}
\title{DMR finder}
\usage{
DMRfinder(
  data,
  chromosome = c(1:22, "X", "Y"),
  mismatches = 3,
  icd = 1000,
  illumina = TRUE
)
}
\arguments{
\item{data}{GRanges-class object, named vector, matrix, or data.frame}

\item{chromosome}{Chromosomes as numeric or "X" or "Y"}

\item{mismatches}{Number of non-significant DMPs allowed (default: 3)}

\item{icd}{Inter-CpG distance (default: 1000bp)}

\item{illumina}{indicator specifying to use CpGs (TRUE) or genomic locations (FALSE) (default: TRUE)}
}
\value{
GRanges-class object containing DMRs
}
\description{
DMR finder
}
\details{
Algorithm to identify differentially methylated regions (DMRs)
from a list of differentially methylated probes (DMPs). This was
developed by Roderick Slieker and is described in "Identification 
and systematic annotation of tissue-specific differentially methylated 
regions using the Illumina 450k array" Slieker et al. (2013)
}
\author{
R Sliecker, E.W Lameijer, and M. van Iterson
}
