% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.gcscatterplot}
\alias{.gcscatterplot}
\title{quality control plot}
\usage{
.gcscatterplot(
  data,
  x,
  y,
  col,
  xlab = NULL,
  ylab = NULL,
  threshold_h = NULL,
  threshold_v = NULL,
  outliers = NULL,
  xlim,
  ylim,
  main
)
}
\arguments{
\item{data}{input data}

\item{x}{x}

\item{y}{y}

\item{col}{col}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{threshold_h}{threshold_h}

\item{threshold_v}{threshold_v}

\item{outliers}{outliers}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{main}{main}
}
\value{
plot
}
\description{
basic GC scatter plot
}
\author{
Tom Kuipers
}
