% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_outliers}
\alias{get_outliers}
\title{get outliers}
\usage{
get_outliers(object, thresholds)
}
\arguments{
\item{object}{input sData}

\item{thresholds}{threshold values}
}
\value{
outliers
}
\description{
get outliers
}
\author{
Tom Kuipers
}
