% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSex.DNAmArray}
\alias{getSex.DNAmArray}
\title{Determine sex based on X chr methylation status}
\usage{
getSex.DNAmArray(beta, cutbeta = c(0.2, 0.6), nx = 3000)
}
\arguments{
\item{beta}{beta matrix}

\item{cutbeta}{beta values threshold use values between [0.2-0.6]}

\item{nx}{number of X chr probes to determine sex is Male default 3000}
}
\value{
sex prediction
}
\description{
Determine sex based on X chr methylation status
}
\author{
rcslieker
}
