% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.GoNLsnps}
\alias{hg19.GoNLsnps}
\title{title}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{maye Zhou here?}
}
\usage{
hg19.GoNLsnps
}
\description{
description
}
\keyword{datasets}
