% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.manifest.pop}
\alias{hm450.manifest.pop}
\title{title}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{}
}
\usage{
hm450.manifest.pop
}
\description{
description
}
\keyword{datasets}
