% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbccPredictor.R
\name{plot_wbcc}
\alias{plot_wbcc}
\title{Validation plots}
\usage{
plot_wbcc(measured, predicted, ...)
}
\arguments{
\item{measured}{Measured cell percentages}

\item{predicted}{Cell percentages predicted by the model}

\item{...}{Additional parameters for plsr see ?plsr}
}
\value{
Correlation plots for measured and predicted cell percentages
}
\description{
Plots to validate predictor a predictor trained using the train_wbcc function
}
\author{
ljsinke
}
