% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBS}
\alias{plotBS}
\title{BS quality control plot}
\usage{
plotBS(object, threshold, col = NULL)
}
\arguments{
\item{object}{input sData}

\item{threshold}{threshold value}

\item{col}{col}
}
\value{
plot
}
\description{
basic BS scatter plot
}
\author{
Tom Kuipers
}
