% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDP}
\alias{plotDP}
\title{DP quality control plot}
\usage{
plotDP(object, threshold, col = NULL)
}
\arguments{
\item{object}{input sData}

\item{threshold}{threshold value}

\item{col}{col}
}
\value{
plot
}
\description{
basic DP scatter plot
}
\author{
Tom Kuipers
}
