% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHC}
\alias{plotHC}
\title{HC quality control plot}
\usage{
plotHC(object, threshold, col = NULL)
}
\arguments{
\item{object}{input sData}

\item{threshold}{threshold value}

\item{col}{col}
}
\value{
plot
}
\description{
basic HC scatter plot
}
\author{
Tom Kuipers
}
