% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMU}
\alias{plotMU}
\title{MU quality control plot}
\usage{
plotMU(object, threshold, col = NULL)
}
\arguments{
\item{object}{input sData}

\item{threshold}{threshold value}

\item{col}{col}
}
\value{
plot
}
\description{
basic MU scatter plot
}
\author{
Tom Kuipers
}
