% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbccPredictor.R
\name{predict_wbcc}
\alias{predict_wbcc}
\title{Predict cell-type composition from methylation values}
\usage{
predict_wbcc(pred, data, covar, transformation = function(x) x,
  ncomp = NULL, impute = TRUE, ...)
}
\arguments{
\item{pred}{A matrix or mvr predictor, trained using train_wbcc}

\item{data}{Matrix of DNA methylation beta values}

\item{covar}{Matrix of covariates to correct for (must match those in predictor)}

\item{transformation}{Transformation to apply to predicted values (default: no transformation)}

\item{ncomp}{Optimal number of components}

\item{impute}{Whether to impute missing values (default: TRUE)}

\item{...}{Additional parameters for plsr see ?plsr}
}
\value{
Predicted cell percentages
}
\description{
Predict cell percentages based on a matrix or mvr-based model
}
\author{
mvaniterson, lsinke
}
