% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{preprocessFunnorm.DNAmArray}
\alias{preprocessFunnorm.DNAmArray}
\title{preprocessFunnormv2}
\usage{
preprocessFunnorm.DNAmArray(rgSet, nPCs = 2, sex = NULL, bgCorr = TRUE,
  dyeCorr = TRUE, verbose = TRUE, keepCN = TRUE)
}
\arguments{
\item{rgSet}{object of class 'RGChannelSet'}

\item{nPCs}{number of PC to use for normalization}

\item{sex}{vector of sexes of each sample}

\item{bgCorr}{logical TRUE/FALSE}

\item{dyeCorr}{logical TRUE/FALSE}

\item{verbose}{logical TRUE/FALSE}

\item{keepCN}{logical TRUE/FALSE}

\item{...}{optional arguments to ratioConvert}
}
\value{
object of class 'GenomicRatioSet'
}
\description{
preprocessFunnorm from minfi with additional return options
}
\details{
preprocessFunnorm from minfi with additional return options: keepCN and what
}
\author{
mvaniterson
}
