% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{probeFiltering}
\alias{probeFiltering}
\title{filter on probe-level: number of beads and zero intensities}
\usage{
probeFiltering(RGset, cutbead = 3, zeroint = TRUE, verbose = TRUE)
}
\arguments{
\item{RGset}{a RGChannelSetExtended}

\item{cutbead}{threshold number of beads}

\item{zeroint}{filter out zero intenseties default is TRUE}

\item{verbose}{Default is TRUE}
}
\value{
RGChannelSet
}
\description{
filter on probe-level: number of beads and zero intensities
}
\author{
mvaniterson
}
