% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeMasking.R
\name{probeMasking}
\alias{probeMasking}
\title{Cross-reactive and Polymorphic Probe Masking}
\usage{
probeMasking(values, array = c("EPIC", "450"), genome = c("hg19",
  "hg38"), verbose = TRUE)
}
\arguments{
\item{values}{beta or M values}

\item{array}{EPIC or 450 array}

\item{genome}{hg19 or hg38 as reference}

\item{verbose}{default is TRUE}
}
\value{
beta or M values
}
\description{
This function masks cross-reactive and polymorphic probes as specified
in Zhou et al. "Comprehensive characterization, annotation and innovative
use of Infinium DNA methylation BeadChip probes" (2017).
}
\details{
Specification for both EPIC and HM450 arrays is possible and masking is
genome dependent; different probes are filtered depending
on whether hg38 or hg19 is specified as a reference.
}
\author{
ljsinke
}
