% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qqpf}
\alias{qqpf}
\title{Printer friendly qq-plot}
\usage{
qqpf(pval, p = 0.9, k = 7, add = FALSE, show.fit = FALSE,
  pch = 16, col = 1, nhighlight = 6, bty = "n", main = "", ...)
}
\arguments{
\item{pval}{vector of p-values}

\item{p}{proportion that should be interpolated}

\item{k}{number of points used to interpolate}

\item{add}{add qq-plot to existing qq-plot}

\item{show.fit}{show the interpolation fit}

\item{pch}{default plot pars}

\item{col}{default plot pars}

\item{nhighlight}{number of highlighted points}

\item{bty}{default plot pars}

\item{main}{default plot pars}

\item{...}{additional graphical parameters}
}
\value{
plot
}
\description{
Printer friendly qq-plot
}
\details{
Use interpolation on the 'uninteresting' P-values to reduce the number
of points that are drawn.
}
\author{
mvaniterson
}
