% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{reduce}
\alias{reduce}
\title{Extract functional normalized data according to filter data}
\usage{
reduce(GRset, RGset, what = c("beta", "M"), cutp = 0.01,
  cutsamples = 0.95, cutcpgs = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{GRset}{'GenomicRatioSet' output from functional normalization}

\item{RGset}{'RGset' output from filtering}

\item{what}{return type either M- or beta-values}

\item{cutp}{detection p-value threshold}

\item{cutsamples}{threshold removing samples}

\item{cutcpgs}{threshold removing probes}

\item{verbose}{Default is TRUE}

\item{...}{optional arguments getBeta}
}
\value{
matrix with  M- or beta-values
}
\description{
Extract functional normalized data according to filter data
}
\details{
Since functional normalization doesn't accept NA e.g. after probe filter the
normalized 'GenomicRatioSet' and filtered 'RGChannelSet' need to be merged.
Additionally if M-values are calculated we set -/+Inf values to +/-16
more details
}
\author{
mvaniterson
}
