% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rotateData}
\alias{rotateData}
\title{rotate data}
\usage{
rotateData(data, columns)
}
\arguments{
\item{data}{input data}

\item{columns}{col}
}
\value{
data
}
\description{
rotate data
}
\author{
Tom Kuipers
}
