% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{screeplot}
\alias{screeplot}
\title{Scree plot of the 450k control probe intensities}
\usage{
screeplot(RGset, nmax = 10)
}
\arguments{
\item{RGset}{object of class 'RGChannelSet'}

\item{nmax}{maximal number of PC's to show}
}
\value{
outut of prcomp using the control probe matrix
}
\description{
Scree plot of the 450k control probe intensities
}
\details{
Scree-plot to detect the optimal number of PC to use in Functional Normalization
}
\author{
mvaniterson
}
