% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbccPredictor.R
\name{train_wbcc}
\alias{train_wbcc}
\title{Train a predictor for cell-type composition}
\usage{
train_wbcc(data, covar, cellPer, model = formula(cellPer ~ covar + data),
  ncomp = 50, keep.model = FALSE, ...)
}
\arguments{
\item{data}{Matrix with DNA methylation beta-values}

\item{covar}{Matrix of covariates to correct for in the model}

\item{cellPer}{Matrix of cell percentages on which the predictor will be trained}

\item{model}{Formula (default: cellPer ~ covar + data)}

\item{ncomp}{Number of PLS components to train (default: 50)}

\item{keep.model}{Logical specifying whether to return the model (default: FALSE)}

\item{...}{Additional parameters for plsr see ?plsr}
}
\value{
Prediction model PLSR object
}
\description{
Train the cell-type composition predictor using partial least squares regression on mehtylation values
}
\author{
mvaniterson
}
