% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{DMRfinder}
\alias{DMRfinder}
\title{DMR finder}
\usage{
DMRfinder(regions, mismatches = 3, icd = 1000)
}
\arguments{
\item{regions}{GRanges-object, named vector or matrix/data.frame}

\item{mismatches}{number of non-significant DMPs default <= 3}

\item{icd}{inter-CpG distance default 1000bp}
}
\value{
GRanges object with DMRs
}
\description{
DMR finder
}
\details{
Algorithm is described in Identification and systematic annotation
of tissue-specific differentially methylated regions using the
Illumina 450k array Slieker et al. Epigentics and Chromatin 2013,
6,26.
}
\examples{
require(FDb.InfiniumMethylation.hg19)
feats <- features(FDb.InfiniumMethylation.hg19)
regions <- feats[seqnames(feats) \%in\% c("chr21", "chr22")]
mcols(regions)  <- NULL ##drop unnecessary metadata
mcols(regions)$crit <- rbinom(length(regions), 1, prob=0.2)

##input GRanges
(dmrsR <- DMRfinder(regions))

##input named-vector
dmps <- mcols(regions)$crit
names(dmps) <- names(regions)
(dmrsV <- DMRfinder(dmps))

##bed-like data.frame
regions <- as.data.frame(regions)
regions <- regions[, c("seqnames", "start", "crit")]
colnames(regions) <- c("chr", "pos", "crit")
(dmrsD <- DMRfinder(regions))

##all equal
length(dmrsR)
length(dmrsV)
length(dmrsD)

head(dmrsR)
head(dmrsV)
head(dmrsD)

dmrsR == dmrsV
dmrsR == dmrsD
dmrsV == dmrsD
}
\author{
R Sliecker, E.W Lameijer and M. van Iterson
}
