% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{screeplot}
\alias{screeplot}
\title{Scree plot of the control probe intensities}
\usage{
screeplot(RGset, nmax = 10)
}
\arguments{
\item{RGset}{object of class 'RGChannelSet'}

\item{nmax}{maximal number of PCs to show}
}
\value{
output of prcomp using the control probe matrix
}
\description{
Scree plot of the control probe intensities (450k or EPIC)
}
\details{
Scree-plot to detect the optimal number of PCs to use in Functional Normalization
}
